0704/송강현/Upload 페이지 컴포넌트 추상화작업으로 인한 복잡도 최적화
